use nr f instead of (float) nr

+++ src/test/java/com/google/sps/PlaceGuideServletTest.java
@@ -109,24 +109,24 @@ public final class PlaceGuideServletTest {
   private static final long C_OUTER_PUBLIC_ID = 67890;
   private static final long C_OUTER_PRIVATE_ID = 9876;
   private static final String CREATOR_C_ID = ID_USER_C;
-  private static final GeoPt C_INNER_PUBLIC_COORDINATE = new GeoPt((float) 10, (float) -5);
-  private static final GeoPt C_INNER_PRIVATE_COORDINATE = new GeoPt((float) -14, (float) 14);
-  private static final GeoPt C_OUTER_PUBLIC_COORDINATE = new GeoPt((float) 5, (float) -20);
-  private static final GeoPt C_OUTER_PRIVATE_COORDINATE = new GeoPt((float) -30, (float) -5);
+  private static final GeoPt C_INNER_PUBLIC_COORDINATE = new GeoPt(10f, -5f);
+  private static final GeoPt C_INNER_PRIVATE_COORDINATE = new GeoPt(-14f, 14f);
+  private static final GeoPt C_OUTER_PUBLIC_COORDINATE = new GeoPt(5f, -20f);
+  private static final GeoPt C_OUTER_PRIVATE_COORDINATE = new GeoPt(-30f, -5f);
   // PlaceGuides of user D.
   private static final long D_INNER_PUBLIC_ID = 567890;
   private static final long D_INNER_PRIVATE_ID = 987650;
   private static final long D_OUTER_PUBLIC_ID = 678900;
   private static final long D_OUTER_PRIVATE_ID = 98760;
   private static final String CREATOR_D_ID = ID_USER_D;
-  private static final GeoPt D_INNER_PUBLIC_COORDINATE = new GeoPt((float) 10, (float) 5);
-  private static final GeoPt D_INNER_PRIVATE_COORDINATE = new GeoPt((float) -14, (float) 14);
-  private static final GeoPt D_OUTER_PUBLIC_COORDINATE = new GeoPt((float) 60, (float) 10);
-  private static final GeoPt D_OUTER_PRIVATE_COORDINATE = new GeoPt((float) -10, (float) -45);
+  private static final GeoPt D_INNER_PUBLIC_COORDINATE = new GeoPt(10f, 5f);
+  private static final GeoPt D_INNER_PRIVATE_COORDINATE = new GeoPt(-14f, 14f);
+  private static final GeoPt D_OUTER_PUBLIC_COORDINATE = new GeoPt(60f, 10f);
+  private static final GeoPt D_OUTER_PRIVATE_COORDINATE = new GeoPt(-10f, -45f);
   // Corners of the rectangle for the queried map area.
   // This is a map area which doesn't cross the IDL.
-  private static final GeoPt NORTH_EAST_CORNER = new GeoPt((float) 15, (float) 15);
-  private static final GeoPt SOUTH_WEST_CORNER = new GeoPt((float) -15, (float) -15);
+  private static final GeoPt NORTH_EAST_CORNER = new GeoPt(15f, 15f);
+  private static final GeoPt SOUTH_WEST_CORNER = new GeoPt(-15f, -15f);
 
   private final PlaceGuide testInnerPublicPlaceGuideC =
       new PlaceGuide.Builder(
